🚚 extract prompt entity and repository for models
