+++ backend/models/__init__.py
@@ -4,11 +4,9 @@ from .brains_subscription_invitations import BrainSubscription
 from .chat import Chat, ChatHistory
 from .chats import ChatMessage, ChatQuestion
 from .files import File
-from .prompt import Prompt, PromptStatusEnum
 from .settings import (BrainRateLimiting, BrainSettings, ContactsSettings,
-                       ResendSettings, get_embeddings,
-                       get_documents_vector_store, get_embeddings,
-                       get_supabase_client, get_supabase_db)
+                       ResendSettings, get_documents_vector_store,
+                       get_embeddings, get_supabase_client, get_supabase_db)
 from .user_usage import UserUsage
 
 # TODO uncomment the below import when start using SQLalchemy
