+++ backend/models/databases/supabase/__init__.py
@@ -1,13 +1,12 @@
+from models.databases.supabase.api_brain_definition import ApiBrainDefinitions
 from models.databases.supabase.api_key_handler import ApiKeyHandler
 from models.databases.supabase.brains import Brain
-from models.databases.supabase.brains_subscription_invitations import BrainSubscription
+from models.databases.supabase.brains_subscription_invitations import \
+    BrainSubscription
 from models.databases.supabase.chats import Chats
 from models.databases.supabase.files import File
 from models.databases.supabase.knowledge import Knowledges
 from models.databases.supabase.notifications import Notifications
 from models.databases.supabase.onboarding import Onboarding
-from models.databases.supabase.prompts import Prompts
 from models.databases.supabase.user_usage import UserUsage
 from models.databases.supabase.vectors import Vector
-
-from models.databases.supabase.api_brain_definition import ApiBrainDefinitions
