+++ backend/models/databases/supabase/supabase.py
@@ -9,7 +9,6 @@ from models.databases.supabase import (
     Knowledges,
     Notifications,
     Onboarding,
-    Prompts,
     UserUsage,
     Vector,
 )
@@ -26,7 +25,6 @@ class SupabaseDB(
     Chats,
     Vector,
     Onboarding,
-    Prompts,
     Notifications,
     Knowledges,
     ApiBrainDefinitions,
@@ -40,7 +38,6 @@ class SupabaseDB(
         ApiKeyHandler.__init__(self, supabase_client)
         Chats.__init__(self, supabase_client)
         Vector.__init__(self, supabase_client)
-        Prompts.__init__(self, supabase_client)
         Notifications.__init__(self, supabase_client)
         Knowledges.__init__(self, supabase_client)
         Onboarding.__init__(self, supabase_client)
