+++ backend/modules/prompt/entity/__init__.py
@@ -1,0 +1,1 @@
+from .prompt import Prompt, PromptStatusEnum, CreatePromptProperties, PromptUpdatableProperties, DeletePromptResponse
