+++ backend/modules/prompt/entity/prompt.py
@@ -1,0 +1,40 @@
+from enum import Enum
+from typing import Optional
+from uuid import UUID
+
+from pydantic import BaseModel
+
+
+class PromptStatusEnum(str, Enum):
+    private = "private"
+    public = "public"
+
+
+class Prompt(BaseModel):
+    title: str
+    content: str
+    status: PromptStatusEnum = PromptStatusEnum.private
+    id: UUID
+
+
+class CreatePromptProperties(BaseModel):
+    """Properties that can be received on prompt creation"""
+
+    title: str
+    content: str
+    status: PromptStatusEnum = PromptStatusEnum.private
+
+
+class PromptUpdatableProperties(BaseModel):
+    """Properties that can be received on prompt update"""
+
+    title: Optional[str]
+    content: Optional[str]
+    status: Optional[PromptStatusEnum]
+
+
+class DeletePromptResponse(BaseModel):
+    """Response when deleting a prompt"""
+
+    status: str = "delete"
+    prompt_id: UUID
